+++ changelog.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog][changelog]
 and this project adheres to [Semantic Versioning][semver].
 
 [//]: ==========================================================================
-## [Unreleased]
+## [0.2.0] - 2018-07-07
+### Added
+- Now avoids to propose the current layout as the default dmenu choice.  
+  This increases ergonomy, as toggling between two layout if way faster now.
+
+### Changed
+- Script is now in python.
 
 [//]: ==========================================================================
 ## [0.1.1] - 2017-11-29
@@ -23,5 +29,6 @@ Initial release.
 [changelog]: http://keepachangelog.com/en/1.0.0/
 [semver]: http://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/mpoquet/dmenu-setxkbmap/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/mpoquet/dmenu-setxkbmap/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/mpoquet/dmenu-setxkbmap/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/mpoquet/dmenu-setxkbmap/compare/v0.1.0...v0.1.1